Made IFileSystem internal and moved to Edge
